+++ ErsatzTV.FFmpeg/Filter/HardwareUploadFilter.cs
@@ -6,7 +6,7 @@ public class HardwareUploadFilter : BaseFilter
 
     public HardwareUploadFilter(FFmpegState ffmpegState) => _ffmpegState = ffmpegState;
 
-    public override string Filter => _ffmpegState.HardwareAccelerationMode switch
+    public override string Filter => _ffmpegState.EncoderHardwareAccelerationMode switch
     {
         HardwareAccelerationMode.None => string.Empty,
         HardwareAccelerationMode.Nvenc => "hwupload_cuda",
@@ -15,7 +15,7 @@ public class HardwareUploadFilter : BaseFilter
         _ => "hwupload"
     };
 
-    public override FrameState NextState(FrameState currentState) => _ffmpegState.HardwareAccelerationMode switch
+    public override FrameState NextState(FrameState currentState) => _ffmpegState.EncoderHardwareAccelerationMode switch
     {
         HardwareAccelerationMode.None => currentState,
         _ => currentState with { FrameDataLocation = FrameDataLocation.Hardware }